requirements.txt
cffi==1.17.1
cryptg==0.5.0.post0
ffmpeg-python==0.2.0
future==1.0.0
numpy==2.1.3
opencv-python==4.10.0.84
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
rsa==4.9
Telethon==1.37.0
tqdm==4.67.0
